Sitemap / Blog post category-aware background.
Gem to enable sitemap/gist (needed for next posts).

Implemented a logic to make background of blog posts being rendered inside an iframe,
and it's content is based on post's first category.
If no background exists for that category, the iframe tag is not rendered.
Currently only ABAP background implemented.